Complete overhaul of Reserve.cs
Added to Email and Phone Number to reservation form.
Made it so that if there is a user logged in, they don't have to fill in Reservation Name, Reservation Email, Reservation Phone Number.
Added a method that checks if the date that the user inputs is in the future, and also doesn't fall on a Sunday (have not fully implemented this yet)