Food menu updated
made a sortByPrice in MenuDis.cs

added all the dishes to Dishes.json

changed DishDetails in Menu.cs so the description is more detailed.

changed it so the menu when u open it is always the month that it is irl.

Made a way to edit a dish in the json file with the method changeDish